Made "frame" a keyword parameter for filters
Filters get the webcam frame as keyword parameter now and must capture
*args and **kwargs, so the program can in the future provide parameters
that are unknown to the filter. *args will only contain arguments for
the filter and kwargs will contain arguments for the filter, that are not
set by the program itself. For example you cannot use a "frame" parameter.