Fixed overlays and result filters when no background is used
Removed early return of the unchanged frame when no virtual background
is used. The check if the frame can be returned would need to be
extended more and more with new features that change the frame and the
program is not needed anyway when you completely disable background
replacement and all filters.