Fix problems when using multiple filters
The filter function, and args/kwargs were not copied in
get_imagefilters, but used a reference when they were used with
parameters, so only the last filter was applied. This is fixed
by using default arguments for the filter_with_parameters inner
function, as default arguments are computed at creation of the function.