Allow foreground filters to change the mask
The mask is now stored in the alpha channel of the foreground frame, so
foreground filters can apply transformations that change the mask.

The settings `flip_horizontal` and `flip_vertical` are now replaced by
the filter `flip`.

All filters were adapted to handle frames with 4 color channels.
For most filters the alpha channel should just be copied, geometric
transformations like `flip` should apply the same transformation to the
mask.