Convert the frame from uint8 to float
Convert the frame from uint8 to float in the range 0-255 in the beginning,
so all filters can assume that they are able to multiply, add substract
without numeric overflow or underflow.

This also improves the performance, as the frame needs to be converted
less often between uint8 and other data types.