Added a video input filter
The "video" filter works like the "image_sequence" filter, but reads
the frames directly from a video with a format that is supported by
opencv. Note that animated GIFs are a supported video format.

To avoid having too many frames in the result sequence, the target_fps
parameter allows to skip frames.
When the video has 30 fps and target_fps is 10, only every 3rd image
is added to the image sequence.