Do not cache the frames of a background video (fixes #71)
The previous implementation of the video filter was made for short
animations, which are loaded one time into an array, which is then used
as a ring buffer. This consumes a lot of memory for longer videos.

The new implementation reads the frames from the video when they are
needed and opens the video file again when the last frame was read.